+++ retailmind-backend/src/main/java/com/retailmind/repository/EventoRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EventoRepository extends JpaRepository<Evento, Long> {
+public interface EventoRepository extends JpaRepository<Evento, Integer> {
 
     Page<Evento> findBySesion_SessionId(String sessionId, Pageable pageable);
 }
